Add confirm to remove

+++ js/app.controller.js
@@ -69,6 +69,8 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    const isRemove = confirm('Are you sure you want to remove this location?')
+    if (!isRemove) return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
